requirements.txt
ipython==8.21.0
openai==1.30.1
pandas==2.2.2
python-dotenv==1.0.1